@@ -1,5 +1,5 @@
 {
-    "name": "mrabbani/laravel-module-manager",
+    "name": "itfixit/laravel-module-manager",
     "description": "Laravel Module Management",
     "type": "laravel",
     "authors": [
@@ -9,9 +9,9 @@
         }
     ],
     "require": {
-        "php": "^7.1.3",
-        "illuminate/console": "5.5.*|5.6.*",
-        "illuminate/database": "5.5.*|5.6.*"
+        "php": "^8.0",
+        "illuminate/console": "^6|^7|^8|^9",
+        "illuminate/database": "^6|^7|^8|^9"
     },
     "autoload": {
         "psr-4": {
@@ -30,6 +30,6 @@
             ]
         }
     },
-
+    "minimum-stability": "dev",
     "license": "MIT"
 }
