@@ -1,7 +1,9 @@
 <?php namespace Mrabbani\ModuleManager\Console\Generators;
 
+use Exception;
 use Illuminate\Console\Command;
 use Illuminate\Database\Migrations\MigrationCreator;
+use Illuminate\Filesystem\Filesystem;
 use Illuminate\Support\Composer;
 
 class MakeMigration extends Command
@@ -28,43 +30,45 @@ class MakeMigration extends Command
      *
      * @var string
      */
-    protected $type = 'Make migration';
+    protected string $type = 'Make migration';
 
     /**
      * The migration creator instance.
      *
-     * @var \Illuminate\Database\Migrations\MigrationCreator
+     * @var MigrationCreator|null
      */
-    protected $creator;
+    protected ?MigrationCreator $creator;
 
     /**
      * The Composer instance.
      *
-     * @var \Illuminate\Support\Composer
+     * @var Composer|null
      */
-    protected $composer;
+    protected ?Composer $composer;
 
     /**
      * Create a new migration install command instance.
      *
-     * @param  \Illuminate\Database\Migrations\MigrationCreator $creator
-     * @param  \Illuminate\Support\Composer $composer
+     * @param Composer         $composer
+     *
      * @return void
      */
-    public function __construct(MigrationCreator $creator, Composer $composer)
+    public function __construct(Composer $composer)
     {
         parent::__construct();
 
-        $this->creator = $creator;
         $this->composer = $composer;
         $this->composer->setWorkingPath(base_path());
     }
 
-    public function handle()
+    /**
+     * @throws Exception
+     */
+    public function handle(): void
     {
         // It's possible for the developer to specify the tables to modify in this
         // schema operation. The developer may also specify if this table needs
-        // to be freshly created so we can create the appropriate migrations.
+        // to be freshly created, so we can create the appropriate migrations.
         $name = trim($this->argument('name'));
 
         $table = $this->option('table');
@@ -86,18 +90,24 @@ class MakeMigration extends Command
     /**
      * Write the migration file to disk.
      *
-     * @param  string $name
-     * @param  string $table
-     * @param  bool $create
-     * @return string
+     * @param        $name
+     * @param string $table
+     * @param bool   $create
+     *
+     * @throws Exception
      */
-    protected function writeMigration($name, $table, $create)
+    protected function writeMigration($name, string $table, bool $create): void
     {
         $path = $this->getMigrationPath();
 
+        $this->creator = new MigrationCreator(
+            new Filesystem(),
+            ''
+        );
+
         $file = pathinfo($this->creator->create($name, $path, $table, $create), PATHINFO_FILENAME);
 
-        return $this->line("<info>Created Migration:</info> {$file}");
+        $this->line("<info>Created Migration:</info> $file");
     }
 
     /**
@@ -105,11 +115,11 @@ class MakeMigration extends Command
      *
      * @return string
      */
-    protected function getMigrationPath()
+    protected function getMigrationPath(): string
     {
         $module = get_module_information($this->argument('alias'));
         $baseDir = get_base_folder(array_get($module, 'file'));
-        $path = $baseDir . 'database/migrations';
-        return $path;
+
+        return $baseDir . 'database/migrations';
     }
 }
