@@ -14,7 +14,9 @@ class BootstrapModuleServiceProvider extends ServiceProvider
     public function boot()
     {
         app()->booted(function () {
-            $this->booted();
+            $this->booted(static function() {
+                
+            });
         });
     }
 
