@@ -18,7 +18,9 @@ class InstallModuleServiceProvider extends ServiceProvider
     public function boot()
     {
         app()->booted(function () {
-            $this->booted();
+            $this->booted(static function() {
+                
+            });
         });
     }
 
